Change URL with given parameters in changeActivePage function
The function to change the URL (changeActivePage) was creating a query string with only one parameter for the section. The app needs to be able to set the category and the product ID in the URL as well.

I have refactored the function so its able to take in more parameters. It also works if only one or two parameters are given.