Break out logic from changeActivePage function
The function changeActivePage was doing two things: it was creating a URL as well as updating the page. The functionality to only create a URL was needed inside a product card in the product list, without reloading a page.

Now the logic is simplified and broken out into two functions.

The new function createURL returns a string with a complete URL, including the given parameters.

changeActivePage function uses createURL() and then it updates the page.