Create addToCart function
The addToCart function saves the selected product in local storage. It creates a new object for the selected product with its ID.

The function first checks if there’s already something in the cart (key in local storage). If there is, it takes the existing products. Then it updates the local storage with existing products and selected new product.

The addToCart function runs when the button to buy a product is clicked. It can also be used inside a single product page.

The function still needs to be able to count number of items. For example if you want to buy more of the same item.