Include complete data and total price for each product
The addToCart function was originally built to save ID and quantity of selected product. This way other features can fetch the latest product data from JSON file as needed, using ID as reference.

Now the addToCart function has been changed so that it saves all data from selected product.
A risk for this method might be that the cart in local storage won't have the updated data from JSON. For example if there is a change in price, after the user has added to cart.

The function now also saves and updates the total price for selected amount of specific product.